move shared to shared.views
