+++ src/edu/wpi/zirconium/lettercraze/builder/LetterCrazeBuilder.java
@@ -2,7 +2,7 @@ package edu.wpi.zirconium.lettercraze.builder;
 
 import com.sun.javafx.application.LauncherImpl;
 import edu.wpi.zirconium.lettercraze.entities.LevelPackData;
-import edu.wpi.zirconium.lettercraze.views.SplashScreen;
+import edu.wpi.zirconium.lettercraze.shared.views.SplashScreen;
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Parent;
