+++ src/edu/wpi/zirconium/lettercraze/builder/views/Builder.fxml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
-<?import edu.wpi.zirconium.lettercraze.shared.BoardView?>
+<?import edu.wpi.zirconium.lettercraze.shared.views.BoardView?>
 <?import javafx.geometry.*?>
 <?import javafx.scene.control.*?>
 <?import javafx.scene.layout.*?>
