+++ src/edu/wpi/zirconium/lettercraze/player/controllers/LevelScreenControllers.java
@@ -6,8 +6,8 @@ import edu.wpi.zirconium.lettercraze.entities.Tile;
 import edu.wpi.zirconium.lettercraze.entities.bindings.WordStringBinding;
 import edu.wpi.zirconium.lettercraze.player.LetterCrazePlayer;
 import edu.wpi.zirconium.lettercraze.player.views.LevelScreen;
-import edu.wpi.zirconium.lettercraze.shared.BoardView;
-import edu.wpi.zirconium.lettercraze.shared.TileView;
+import edu.wpi.zirconium.lettercraze.shared.views.BoardView;
+import edu.wpi.zirconium.lettercraze.shared.views.TileView;
 import edu.wpi.zirconium.lettercraze.utils.ContainsBinding;
 import javafx.event.EventHandler;
 import javafx.fxml.FXML;
