+++ src/edu/wpi/zirconium/lettercraze/player/views/Level.fxml
@@ -1,11 +1,13 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
-<?import edu.wpi.zirconium.lettercraze.shared.BoardView?>
-<?import javafx.geometry.*?>
+<?import edu.wpi.zirconium.lettercraze.shared.views.BoardView?>
+<?import javafx.geometry.Insets?>
 <?import javafx.scene.control.Button?>
 <?import javafx.scene.control.TextArea?>
-<?import javafx.scene.Group?>
-<?import javafx.scene.layout.*?>
+<?import javafx.scene.layout.BorderPane?>
+<?import javafx.scene.layout.HBox?>
+<?import javafx.scene.layout.StackPane?>
+<?import javafx.scene.layout.VBox?>
 <?import javafx.scene.shape.Rectangle?>
 <?import javafx.scene.shape.SVGPath?>
 <?import javafx.scene.text.Font?>
@@ -15,8 +17,8 @@
 <fx:root fx:id="root" type="LevelScreen"
          prefHeight="712.0" prefWidth="1024.0"
          stylesheets="@level.css"
-         fx:controller="edu.wpi.zirconium.lettercraze.player.controllers.LevelScreenControllers"
-         xmlns="http://javafx.com/javafx/8.0.65" xmlns:fx="http://javafx.com/fxml/1">
+         xmlns="http://javafx.com/javafx/8.0.45" xmlns:fx="http://javafx.com/fxml/1"
+         fx:controller="edu.wpi.zirconium.lettercraze.player.controllers.LevelScreenControllers">
    <top>
       <StackPane maxHeight="-Infinity" maxWidth="-Infinity" BorderPane.alignment="TOP_LEFT">
          <BorderPane.margin>
